Function whitenoise_ffi::compute_privacy_usage

#[no_mangle]pub extern "C" fn compute_privacy_usage(
    request_ptr: *const u8, 
    request_length: i32
) -> ByteBuffer

FFI wrapper for compute_privacy_usage

Arguments

request_ptr - a pointer to an array containing the serialized protobuf of RequestComputePrivacyUsage

request_length - the length of the array

Returns

a ByteBufferValidator struct containing a pointer to and length of the serialized protobuf of proto::ResponseComputePrivacyUsage